Refactor query to use index automatically

+++ database/migrations/2023_12_26_110502_create_imports_history_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->string('document_line_number');
             $table->uuid('import_id')->index();
             $table->json('tags')->nullable();
-            $table->tinyInteger('status');
+            $table->smallInteger('status');
             $table->index(['import_id', 'status']);
         });
     }
